+++ inst/shiny-examples/ShinyItemAnalysis/server/Data.R
@@ -14,11 +14,11 @@ checkDataText_Input <- eventReactive(input$submitButton, {
     error_key <- ""
     if (length(key()) == 1) {
       if (key() == "missing") {
-        error_key <- "The key need to be provided!"
+        error_key <- "The key needs to be provided!"
       }
     } else {
       if (ncol(nominal()) != length(key())) {
-        error_key <- "The length of key need to be the same as number of columns of the main dataset!"
+        error_key <- "The length of key needs to be the same as number of columns of the main dataset!"
         error_setting <- T
       }
     }
@@ -29,7 +29,7 @@ checkDataText_Input <- eventReactive(input$submitButton, {
       warni_group <- "The group variable is not provided! DIF and DDF analyses are not available!"
     } else {
       if (nrow(nominal()) != length(group())) {
-        error_group <- "The length of group need to be the same as number of observations in the main dataset!"
+        error_group <- "The length of group needs to be the same as number of observations in the main dataset!"
         error_setting <- T
       }
     }
@@ -40,7 +40,7 @@ checkDataText_Input <- eventReactive(input$submitButton, {
       warni_criterion_variable <- "The criterion variable is not provided! Predictive validity analysis is not available!"
     } else {
       if (nrow(nominal()) != length(criterion())) {
-        error_criterion_variable <- "The length of criterion variable need to be the same as number of observations in the main dataset!"
+        error_criterion_variable <- "The length of criterion variable needs to be the same as number of observations in the main dataset!"
         error_setting <- T
       }
     }
@@ -51,7 +51,7 @@ checkDataText_Input <- eventReactive(input$submitButton, {
       warni_DIFmatching_variable <- "The DIF matching variable is not provided! DIF analyses will use total scores!"
     } else {
       if (nrow(nominal()) != length(DIFmatching())) {
-        error_DIFmatching_variable <- "The length of DIF matching variable need to be the same as number of observations in the main dataset!"
+        error_DIFmatching_variable <- "The length of DIF matching variable needs to be the same as number of observations in the main dataset!"
         error_setting <- T
       }
     }
@@ -334,56 +334,90 @@ output$renderdeleteButtonGroup <- renderUI({
 data_description_Input <- reactive({
   data_name <- input$dataSelect
   txt <- switch(data_name,
-    GMAT_difNLR = "<code>GMAT</code> <a href='https://doi.org/10.1187/cbe.16-10-0307' target='_blank'>
-                (Martinkova et al., 2017) </a> is generated dataset based on parameters of real Graduate Management
-                Admission Test (GMAT; Kingston et al., 1985). However, first two items were simulated to function
-                differently in uniform and non-uniform way respectively. The dataset represents responses of 2,000 subjects
-                (1,000 males, 1,000 females) to multiple-choice test of 20 items. The distribution of total scores is the
-                same for both groups. See <a href='https://doi.org/10.1187/cbe.16-10-0307' target='_blank'> Martinkova et al.
-                (2017)</a> for further discussion. <code>GMAT</code> also containts simulated continuous criterion variable. ",
-    GMAT2_difNLR = "<code>GMAT2</code> (Drabinova & Martinkova, 2017) is simulated dataset based on parameters
-                of real Graduate Management Admission Test (GMAT; Kingston et al., 1985) from <code>difNLR</code> R package .
-                First two items were simulated to function differently in uniform and non-uniform way respectively.
-                The dataset represents responses of 1,000 subjects (500 males, 500 females) to multiple-choice test of 20
-                items.",
-    MSATB_difNLR = "<code>MSAT-B</code> (Drabinova & Martinkova, 2017) is a subset of real Medical School
-                Admission Test in Biology (MSAT-B) in Czech Republic from <code>difNLR</code> R package. The dataset
-                represents responses of 1,407 subjects (484 males, 923 females) to multiple-choice test of 20 items.
-                First item was previously detected as functioning differently. For more details of item selection see
-                Drabinova and Martinkova (2017).",
-    dataMedical_ShinyItemAnalysis = "<code>Medical 100</code> is a real <code>dataMedical</code> dataset of admission test to medical
-                school from <code>ShinyItemAnalysis</code> R package. The data set represents responses of 2,392 subjects
-                (750 males, 1,633 females and 9 subjects without gender specification) to multiple-choice test of 100
-                items.  This dataset contains criterion variable - indicator whether student studies standardly
-                or not.",
-    dataMedicalgraded_ShinyItemAnalysis = "<code>Medical 100 Graded</code> is a real <code>dataMedicalgraded</code> dataset of admission test to medical
-                school from <code>ShinyItemAnalysis</code> R package. The data set represents responses of 2,392 subjects
-                (750 males, 1,633 females and 9 subjects without gender specification) to multiple-choice test of 100
-                items. Each item is graded with 0 to 4 points. Maximum of 4 points were set if all correct answers and none of incorrect answers were selected. This dataset contains criterion variable - indicator whether student studies standardly
-                or not.",
-    HCI_ShinyItemAnalysis = "<code>HCI</code> (McFarland et al., 2017) is a real dataset of Homeostasis
-                Concept Inventory (HCI) from <code>ShinyItemAnalysis</code> R package. The dataset represents responses of
-                651 subjects (405 males, 246 females) to multiple-choice test of 20 items. <code>HCI</code> contains
-                criterion variable -  indicator whether student plans to major in the life sciences.",
-    Science_mirt = "<code>Science</code> dataset represents responses of 392 subjects on a 4-item test describing attitude to science and technology from <code>mirt/ltm</code> R package.
-				        Selected items are <code>Comfort</code>, <code>Work</code>, <code>Future</code>, and <code>Benefit</code>. All items are measured on the
-				        same scale with response categories: <code>'strongly disagree'</code>, <code>'disagree to some extent'</code>, <code>'agree to some extent'</code>,
-				        and <code>'strongly agree'</code>. See Bartholomew et al. (2002) for more details.",
-    LearningToLearn_ShinyItemAnalysis_6 = "<code>Learning To Learn 6</code> is a subset of longitudinal <code>LearningToLearn</code> dataset from
-				        <code>ShinyItemAnalysis</code> R package. It consists of answers to the Learning to Learn test in Grade 6 only. The same respondents were also
-                tested in Grade 9 &ndash; respective data is available in <code>Learning To Learn 9</code> dataset.
-				        The dataset represents binary-coded responses of 782 subjects (391 from basic school track, BS, coded here as 0; 391 from selecive academic track, AS,
-                coded here as 1) to (mostly) multiple-choice test consisting of 41 items within 7 subscales. This dataset was created using propensity score matching
-                algorithm to achieve similar characteristics in both tracks. For further details see <a href='https://doi.org/10.1016/j.learninstruc.2019.101286' target='_blank'>
-				        Martinkova, Hladka, and Potuznikova (2020)</a>.",
-    LearningToLearn_ShinyItemAnalysis_9 = "<code>Learning To Learn 9</code> is a subset of longitudinal <code>LearningToLearn</code> dataset from
-				        <code>ShinyItemAnalysis</code> R package. It consists of answers to the Learning to Learn test in Grade 9 only. The same respondents were also
-                tested in Grade 6 &ndash; respective data is available in <code>Learning To Learn 6</code> dataset.
-				        The dataset represents binary-coded responses of 782 subjects (391 from basic school track, BS, coded here as 0; 391 from selecive academic track, AS,
-                coded here as 1) to (mostly) multiple-choice test consisting of 41 items within 7 subscales. This dataset was created using propensity score matching
-                algorithm to achieve similar characteristics in both tracks. Matching variable for DIF analysis is a total score gained in Grade 6.
-                For further details see <a href='https://doi.org/10.1016/j.learninstruc.2019.101286' target='_blank'>
-				        Martinkova, Hladka, and Potuznikova (2020)</a>."
+    GMAT_difNLR = "<code>GMAT</code> <a href='https://doi.org/10.1187/cbe.16-10-0307' target='_blank'>(Martinkova et al., 2017)</a>
+                is a generated dataset based on parameters of real Graduate Management Admission Test (GMAT; Kingston et al., 1985)
+                from the <code>difNLR</code> package. The dataset represents responses of 2,000 subjects (1,000 males coded as
+                <code>'0'</code>, 1,000 females coded as <code>'1'</code>) to a multiple-choice test of 20 items. It also contains
+                generated continuous criterion intended to be predicted by the test. As discussed in
+                <a href='https://doi.org/10.1187/cbe.16-10-0307' target='_blank'> Martinkova et al. (2017)</a>, the dataset was simulated
+                in order to provide an example of DIF items present even in case of identical total score distributions. To replicate the
+                example provided in <a href='https://doi.org/10.1187/cbe.16-10-0307' target='_blank'> Martinkova et al. (2017)</a>,
+                select the <code>GMAT</code> dataset and go to <code>DIF/Fairness</code> section. ",
+    GMAT2_difNLR = "<code>GMAT2</code> <a href='https://doi.org/10.1111/jedm.12158' target='_blank'>(Drabinova & Martinkova, 2017)</a> is
+                a simulated dataset based on parameters of real Graduate Management Admission Test (GMAT; Kingston et al., 1985) from the
+                <code>difNLR</code> package. First two items were simulated to function differently in uniform and non-uniform way respectively.
+                The dataset represents responses of 1,000 subjects (500 males coded as <code>'0'</code>, 500 females coded as <code>'1'</code>)
+                to a multiple-choice test of 20 items. ",
+    MSATB_difNLR = "<code>MSAT-B</code> <a href='https://doi.org/10.1111/jedm.12158' target='_blank'>(Drabinova & Martinkova, 2017)</a> is
+                a subset of real Medical School Admission Test in Biology (MSAT-B) in Czech Republic from the <code>difNLR</code> package.
+                The dataset represents responses of 1,407 subjects (484 males coded as <code>'0'</code>, 923 females coded as <code>'1'</code>)
+                to a multiple-choice test of 20 items. The first item was previously detected as functioning differently. For more details
+                on item selection see <a href='https://doi.org/10.1111/jedm.12158' target='_blank'>Drabinova and Martinkova (2017)</a>.
+                To replicate the example provided in <a href='https://doi.org/10.1111/jedm.12158' target='_blank'> Drabinova and Martinkova
+                (2017)</a>, select the <code>MSAT-B</code> dataset and go to <code>DIF/Fairness</code> section. ",
+    dataMedical_ShinyItemAnalysis = "<code>Medical 100</code> <a href='https://doi.org/10.5817/TF2017-9-129' target='_blank'>
+                (Martinkova et al., 2017)</a> is a real <code>dataMedical</code> dataset of an admission test to medical
+                school from the <code>ShinyItemAnalysis</code> package. The data set represents responses of 2,392 subjects
+                (750 males coded as <code>'0'</code>, 1,633 females coded as <code>'1'</code>, and 9 subjects without gender
+                specification coded as <code>'NA'</code>) to a multiple-choice test of 100 items. The dataset contains criterion
+                variable &ndash; indicator whether student studies standardly or not. In <a href='https://doi.org/10.5817/TF2017-9-129'
+                target='_blank'> Martinkova et al. (2017)</a>, the dataset was used to demonstrate functionalities of earlier
+                version of the <code>ShinyItemAnalysis</code> interactive app. ",
+    dataMedicalgraded_ShinyItemAnalysis = "<code>Medical 100 Graded</code> is a real <code>dataMedicalgraded</code> dataset
+                of an admission test to medical school from the <code>ShinyItemAnalysis</code> package. The dataset represents
+                responses of 2,392 subjects (750 males coded as <code>'0'</code>, 1,633 females coded as <code>'1'</code>, and
+                9 subjects without gender specification coded as <code>'NA'</code>) to a multiple-choice test of 100 items. Each
+                item is graded with 0 to 4 points. Maximum of 4 points were set if all correct answers and none of incorrect
+                answers were selected. This dataset contains criterion variable &ndash; indicator whether student studies standardly
+                or not. For analyses where dichotomous items are necessary (e.g., logistic models in Regression, IRT models, or
+                DIF detection methods) binarization of the data is applied &ndash; <code>'1'</code> means that student gained maximum
+                number of points, i.e., 4; otherwise the item is scored as <code>'0'</code>. ",
+    HCI_ShinyItemAnalysis = "<code>HCI</code> <a href='http://dx.doi.org/10.1187/cbe.16-10-0305' target='_blank'>(McFarland et al.,
+                2017)</a> is a real dataset of Homeostasis Concept Inventory (HCI) from the <code>ShinyItemAnalysis</code> package.
+                The dataset represents responses of 651 subjects (405 males coded as <code>'0'</code>, 246 females coded as
+                <code>'0'</code>) to a multiple-choice test of 20 items. <code>HCI</code> contains criterion variable &ndash; indicator
+                whether student plans to major in the life sciences. In <a href='http://dx.doi.org/10.1187/cbe.16-10-0305'
+                target='_blank'> McFarland et al. (2017)</a>, the dataset was used to perform complex validation of the HCI
+                assessment instrument. The dataset was also used for demonstrations of the <code>ShinyItemAnalysis</code> package
+                and earlier version of this online app in The R Journal paper by <a href='https://doi.org/10.32614/RJ-2018-074'
+                target='_blank'> Martinkova and Hladka (2018)</a>. ",
+    Science_mirt = "<code>Science</code> dataset represents responses of 392 subjects on a 4-item test describing attitude to science
+                and technology from the <code>mirt/ltm</code> packages. Selected items are <code>Comfort</code>, <code>Work</code>,
+                <code>Future</code>, and <code>Benefit</code>. All items are measured on the same scale with response categories:
+                <code>'strongly disagree'</code>, <code>'disagree to some extent'</code>, <code>'agree to some extent'</code>, and
+                <code>'strongly agree'</code>. See Karlheinz and Melich (1992) for more details. For analyses where dichotomous items
+                are necessary (e.g., logistic models in Regression, IRT models, or DIF detection methods) binarization of the data is
+                applied &ndash; <code>'1'</code> means that respondent <code>'strongly agrees'</code> with given item; otherwise
+                the item is scored <code>'0'</code>. ",
+    LearningToLearn_ShinyItemAnalysis_6 = "<code>Learning To Learn 6</code> <a href='https://doi.org/10.1016/j.learninstruc.2019.101286'
+                target='_blank'> (Martinkova et al., 2020)</a> is a subset of longitudinal <code>LearningToLearn</code> dataset from the
+                <code>ShinyItemAnalysis</code> package. It consists of answers to the Learning to Learn test in Grade 6 only. The same
+                respondents were also tested in Grade 9 &ndash; respective data is available in the <code>Learning To Learn 9</code>
+                dataset. The dataset represents binary-coded responses of 782 subjects (391 from basic school track, BS, coded here as
+                <code>'0'</code>; 391 from selecive academic track, AS, coded here as <code>'1'</code>) to (mostly) multiple-choice test
+                consisting of 41 items within 7 subscales. This dataset was created using propensity score matching algorithm to achieve
+                similar characteristics in both tracks. For further details see <a href='https://doi.org/10.1016/j.learninstruc.2019.101286'
+                target='_blank'> Martinkova, Hladka, and Potuznikova (2020)</a>.",
+    LearningToLearn_ShinyItemAnalysis_9 = "<code>Learning To Learn 9</code> <a href='https://doi.org/10.1016/j.learninstruc.2019.101286'
+                target='_blank'> (Martinkova et al., 2020)</a> is a subset of longitudinal <code>LearningToLearn</code> dataset from the
+                <code>ShinyItemAnalysis</code> package. It consists of answers to the Learning to Learn test in Grade 9 only. The same
+                respondents were also tested in Grade 6 &ndash; respective data is available in <code>Learning To Learn 6</code> dataset.
+				        The dataset represents binary-coded responses of 782 subjects (391 from basic school track, BS, coded here as <code>'0'</code>;
+				        391 from selecive academic track, AS, coded here as <code>'1'</code>) to (mostly) multiple-choice test consisting of 41
+				        items within 7 subscales. The dataset was created using propensity score matching algorithm to achieve similar
+				        characteristics in both tracks. Matching variable for the DIF analysis is the total score gained in Grade 6. In their paper,
+				        <a href='https://doi.org/10.1016/j.learninstruc.2019.101286' target='_blank'> Martinkova et al. (2020)</a> found significant
+				        DIF in change present in some of the items, while no significant difference in change was present in total scores. The dataset
+				        was also used for demonstration purposes in The R Journal paper on the <code>difNLR</code> package
+                <a href='https://doi.org/10.32614/RJ-2020-014' target='_blank'>(Hladka & Martinkova, 2020)</a>. ",
+    AIBS_ShinyItemAnalysis = "<code>AIBS Grant Peer Review Scoring</code> dataset <a href='https://doi.org/10.6084/m9.figshare.12728087'
+                target='_blank'> (Gallo, 2021)</a> comes from a scientific peer review of biomedical applications from a collaborative
+                biomedical research funding program (2014&ndash2017). The data presented in this app include anonymized proposal ID, reviewer ID,
+                and an overall score from three reviewers, scored on a scale from 1.0 (best) to 5.0 (worst) with a 0.1 gradation. The
+                dataset was used by <a href='https://rss.onlinelibrary.wiley.com/loi/1467985x' target='_blank'>Erosheva, Martinkova, and Lee</a>
+                to demonstrate issues with estimation of inter-rater reliability (IRR) in range-restricted samples. To try interactively
+                the range-restricted IRR analysis, select the <code>AIBS</code> dataset and go to the <code>Reliability/Restricted range</code>
+                section."
   )
   txt
 })
@@ -412,118 +446,101 @@ output$data_description <- renderUI({
 # BASIC SUMMARY ####
 
 # * Main dataset ####
+
 # ** Dimension ####
 output$data_rawdata_dim <- renderText({
-  txt <- paste0(
-    "Dataset consists of ", nrow(nominal()),
-    " observations on the ", ncol(nominal()), " items. "
-  )
+  if (input$dataSelect == "AIBS_ShinyItemAnalysis") {
+    txt <- paste0("The AIBS is not a multi-item dataset. It consists of ", nrow(continuous()),
+      " observations on the ", ncol(continuous()), " variables. "
+    )
+  } else {
+    txt <- paste0(
+      "The main dataset consists of ", nrow(continuous()),
+      " observations on the ", ncol(continuous()), " ", data_type(), " items. "
+    )
+  }
+  txt
 })
 
-# ** Summary ####
-# *** Ordinal data ####
-data_ordinal_summary_Input <- reactive({
+# ** Ordinal data summary ####
+data_ordinal_summary <- reactive({
   data_table <- ordinal()
 
-  if (input$data_type == "ordinal" | input$dataSelect == "dataMedicalgraded_ShinyItemAnalysis") {
+  if (data_type() == "ordinal") {
     data_table <- as.data.frame(sapply(data_table, as.numeric))
-    if (input$data_type == "ordinal") {
-      key <- key()
-      # Distinction of MIN and MAX key will be added based on guture submitBUtton changes
-      # ...minKey=keyOrdinalMin
-      # ...maxKey=keyOrdinalMax
-    }
+    key <- key()
 
-    if (input$dataSelect == "dataMedicalgraded_ShinyItemAnalysis") {
-      key <- key()
-      # Distinction of MIN and MAX key will be added
-    }
     data_table_summary <- data.table(
       item_names(),
-      sapply(data_table, min),
-      sapply(data_table, median),
-      sapply(data_table, mean),
-      sapply(data_table, max),
-      sapply(data_table, sd),
+      sapply(data_table, min, na.rm = TRUE),
+      sapply(data_table, median, na.rm = TRUE),
+      sapply(data_table, mean, na.rm = TRUE),
+      sapply(data_table, max, na.rm = TRUE),
+      sapply(data_table, sd, na.rm = TRUE),
       as.numeric(key)
     )
     rownames(data_table_summary) <- item_names()
     colnames(data_table_summary) <- c("Name", "Min", "Median", "Mean", "Max", "SD", "Cut")
     data_table_summary
   } else {
-    # data_table <- sapply(data_table, as.factor)
     colnames(data_table) <- item_names()
     summary(data_table)
   }
 })
 
-output$data_rawdata_summary <- renderPrint({
-  if (dataset$data_type == "ordinal") {
-    data_ordinal_summary_Input()
-  } else {
-    if (dataset$data_type == "nominal") {
-      data_nominal_summary_Input()
-    } else {
-      data_binary_summary_Input()
-    }
-  }
-})
-
-
-# *** Nominal data ####
-data_nominal_summary_Input <- reactive({
+# ** Nominal data summary ####
+data_nominal_summary <- reactive({
   data_table <- nominal()
-  # data_table <- sapply(data_table, as.factor)
   colnames(data_table) <- item_names()
   summary(data_table)
 })
 
-output$data_nominal_summary <- renderPrint({
-  if (input$data_type == "nominal") {
-    data_nominal_summary_Input()
-  } else {
-    ""
-  }
-})
-
-# *** Binary data ####
-data_binary_summary_Input <- reactive({
+# ** Binary data summary ####
+data_binary_summary <- reactive({
   data_table <- binary()
-  # data_table <- sapply(data_table, as.factor)
   colnames(data_table) <- item_names()
   summary(data_table)
 })
 
 output$data_binary_summary <- renderPrint({
-  data_binary_summary_Input()
+  data_binary_summary()
+})
+
+# ** Continuous data summary ####
+data_continuous_summary <- reactive({
+  data_table <- continuous()
+  summary(data_table)
+})
+
+# ** Main dataset summary ####
+output$data_rawdata_summary <- renderPrint({
+  if (dataset$data_type == "ordinal") {
+    data_ordinal_summary()
+  } else if (dataset$data_type == "nominal") {
+    data_nominal_summary()
+  } else if (dataset$data_type == "continuous") {
+    data_continuous_summary()
+  } else {
+    data_binary_summary
+  }
 })
 
 # * Group ####
-data_group_summary_Input <- reactive({
+output$data_group_summary <- renderPrint({
   gr <- as.factor(group())
   summary(gr)
 })
 
-output$data_group_summary <- renderPrint({
-  data_group_summary_Input()
-})
-
 # * Criterion ####
-data_criterion_summary_Input <- reactive({
-  summary(criterion())
-})
-
 output$data_criterion_summary <- renderPrint({
-  data_criterion_summary_Input()
+  criterion <- criterion() # has to be called ousite S3 methods, otherwise returns ugly warnigs
+  summary(criterion)
 })
 
 # * DIF matching ####
-data_DIFmatching_summary_Input <- reactive({
-  summary(DIFmatching())
-})
-
 output$data_DIFmatching_summary <- renderPrint({
-  data_DIFmatching_summary_Input()
+  DIFmatching <- DIFmatching()
+  summary(DIFmatching)
 })
 
 
@@ -548,14 +565,14 @@ output$data_DIFmatching_summary <- renderPrint({
 
 # DATA EXPLORATION ####
 
-# * Main dataset ######
+# * Main dataset ####
 output$headdata <- DT::renderDataTable(
   {
     data_table <- nominal()
     colnames(data_table) <- item_names()
     data_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,
@@ -566,14 +583,15 @@ output$headdata <- DT::renderDataTable(
   )
 )
 
-# * Key ######
+# * Key ####
 output$key <- DT::renderDataTable(
   {
-    key_table <- as.data.table(t(key()))
+    key_table <- key()
+    key_table <- as.data.table(t(key_table))
     colnames(key_table) <- item_names()
     key_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,
@@ -583,15 +601,15 @@ output$key <- DT::renderDataTable(
   )
 )
 
-# * Binary data with total score ######
+# * Binary data with total score ####
 output$sc01 <- DT::renderDataTable(
   {
-    scored_table <- data.table(binary(), total_score())
+    data_table <- binary()
+    scored_table <- data.table(data_table, total_score())
     colnames(scored_table) <- c(item_names(), "Score")
-
     scored_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,
@@ -602,14 +620,15 @@ output$sc01 <- DT::renderDataTable(
   )
 )
 
-# GROUP CONTROL ######
+# * Group ####
 output$group <- DT::renderDataTable(
   {
-    group_table <- t(group())
+    group_table <- group()
+    group_table <- t(group_table)
     colnames(group_table) <- 1:ncol(group_table)
     group_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,
@@ -619,14 +638,15 @@ output$group <- DT::renderDataTable(
   )
 )
 
-# CRITERION VARIABLE CONTROL ######
+# * Criterion variable ####
 output$critvar <- DT::renderDataTable(
   {
-    critvar_table <- t(criterion())
+    critvar_table <- criterion()
+    critvar_table <- t(critvar_table)
     colnames(critvar_table) <- 1:ncol(critvar_table)
     critvar_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,
@@ -636,14 +656,15 @@ output$critvar <- DT::renderDataTable(
   )
 )
 
-# DIF MATCHING VARIABLE CONTROL ######
+# * DIF matching variable ####
 output$difvar <- DT::renderDataTable(
   {
-    difvar_table <- t(DIFmatching())
+    difvar_table <- DIFmatching()
+    difvar_table <- t(difvar_table)
     colnames(difvar_table) <- 1:ncol(difvar_table)
     difvar_table
   },
-  rownames = F,
+  rownames = FALSE,
   style = "bootstrap",
   options = list(
     scrollX = TRUE,