+++ NEWS.md
@@ -13,6 +13,8 @@
   * In the Regression section of the app, missplaced SE for intercepts and slopes were fixed.
 
 #### MAJOR UPDATES
+  * Continuous items are newly allowed, rater variable is added, data upload and warning
+    messages were updated
   * New function ICCrestr() for computing ICC in restricted samples was added.
   * New AIBS dataset was added to the package
   * Functions gDiscrim(), DDplot() and ItemAnalysis were updated, variable names were unified